+++ Pandora/Paradigm/Inventory/Imprint.hs
@@ -8,9 +8,8 @@ import Pandora.Pattern.Functor.Contravariant (Contravariant ((>-|-)))
 import Pandora.Pattern.Functor.Distributive (Distributive ((-<<)))
 import Pandora.Pattern.Functor.Extendable (Extendable ((<<=)))
 import Pandora.Pattern.Functor.Comonad (Comonad)
-import Pandora.Pattern.Functor.Divariant (Divariant ((>->)))
 import Pandora.Pattern.Object.Semigroup (Semigroup ((+)))
-import Pandora.Paradigm.Primary.Algebraic.Exponential ()
+import Pandora.Paradigm.Primary.Algebraic ()
 import Pandora.Pattern.Morphism.Flip (Flip (Flip))
 import Pandora.Paradigm.Controlflow.Effect.Interpreted (Schematic, Interpreted (Primary, run, unite, (||=), (!)))
 import Pandora.Paradigm.Controlflow.Effect.Adaptable (Adaptable)
@@ -26,9 +25,6 @@ instance Contravariant (->) (->) (Flip Imprint a) where
 
 instance Distributive (->) (->) (Imprint e) where
 	f -<< g = Imprint ! (run @(->) <-|- f) -<< g
-
-instance Divariant (->) (->) (->) Imprint where
-	(>->) ab cd bc = ab >-> cd ||= bc
 
 instance Semigroup e => Extendable (->) (Imprint e) where
 	f <<= Imprint x = Imprint ! \e -> f . Imprint ! x . (e +)