+++ Pandora/Paradigm/Primary/Functor/Endo.hs
@@ -3,11 +3,12 @@ module Pandora.Paradigm.Primary.Functor.Endo where
 import Pandora.Pattern.Semigroupoid ((.))
 import Pandora.Pattern.Category (identity, (#))
 import Pandora.Pattern.Functor.Invariant (Invariant ((<!<)))
-import Pandora.Pattern.Functor.Divariant ((>->))
 import Pandora.Pattern.Object.Semigroup (Semigroup ((+)))
 import Pandora.Pattern.Object.Monoid (Monoid (zero))
 import Pandora.Paradigm.Controlflow.Effect.Interpreted (Interpreted (Primary, run, unite, (||=)))
 import Pandora.Paradigm.Primary.Algebraic.Exponential ()
+import Pandora.Paradigm.Primary.Algebraic.Product ((:*:) ((:*:)))
+import Pandora.Paradigm.Primary.Algebraic ((>-|-<-|-))
 
 newtype Endo a = Endo { endo :: a -> a }
 
@@ -17,7 +18,7 @@ instance Interpreted (->) Endo where
 	unite = Endo
 
 instance Invariant Endo where
-	f <!< g = (g >-> f ||=)
+	f <!< g = ((g :*: f >-|-<-|-) ||=)
 
 instance Semigroup (Endo a) where
 	Endo f + Endo g = Endo # g . f