+++ Pandora/Pattern/Functor.hs
@@ -1,6 +1,5 @@
 module Pandora.Pattern.Functor (module Exports) where
 
-import Pandora.Pattern.Functor.Divariant as Exports
 import Pandora.Pattern.Functor.Comonad as Exports
 import Pandora.Pattern.Functor.Monad as Exports
 import Pandora.Pattern.Functor.Representable as Exports